Changed rate limit handler to sleep for a max of 10 seconds.
I'm hitting MULTIPLE rate limits. ~5000 on one request, and 30 in the
next. I don't know if this is some sort of round-robin/big-ip thing or
what, but waiting 2700 seconds when the next request would have worked
anyways seems like a waste.

Also, it seems like the .netrc oauth token is NOT being used. I need
to dig into that.

Signed-off-by: Ryan Davis <zenspider@chef.io>